+++ packages/guards.integer/src/lib/guardInteger.ts
@@ -1,22 +1,23 @@
 import { ZeroDepErrorGuardRange, ZeroDepErrorGuardType } from '@zerodep/guards.errors';
+import { isInteger } from '@zerodep/is.integer';
 
-export interface IGuardIntegerOptions {
+export interface GuardIntegerOptions {
   min?: number;
   max?: number;
 }
 
 // default options
-const defaultOptions: IGuardIntegerOptions = {
+const defaultOptions: GuardIntegerOptions = {
   min: undefined,
   max: undefined,
 };
 
-export const guardInteger = (options: IGuardIntegerOptions = {}) => {
-  const config: IGuardIntegerOptions = { ...defaultOptions, ...options };
+export const guardInteger = (options: GuardIntegerOptions = {}) => {
+  const config: GuardIntegerOptions = { ...defaultOptions, ...options };
 
   return (value: any): number => {
     // we need to check for the typeof first as "undefined" will cause isInteger() to error
-    if (typeof value !== 'number' || !Number.isInteger(value)) {
+    if (!isInteger(value)) {
       const error = new ZeroDepErrorGuardType('Value is not an integer');
       error.value = value;
       throw error;