+++ packages/is.bigint/src/lib/isBigint.ts
@@ -1,0 +1,3 @@
+export const isBigint = (value: any): boolean => {
+  return typeof value === 'bigint';
+};