+++ packages/is.date/src/lib/isDate.ts
@@ -1,0 +1,3 @@
+export const isDate = (value: any): boolean => {
+  return typeof value === 'object' && value instanceof Date;
+};