+++ packages/is.error/src/lib/isError.ts
@@ -1,0 +1,3 @@
+export const isError = (value: any): boolean => {
+  return value instanceof Error;
+};