+++ packages/is.float/src/lib/isFloat.ts
@@ -1,0 +1,8 @@
+export const isFloat = (value: any): boolean => {
+  return (
+    typeof value === 'number' &&
+    !Number.isNaN(value) &&
+    !Number.isInteger(value) &&
+    Number.isFinite(value)
+  );
+};