+++ packages/is.integer/src/lib/isInteger.ts
@@ -1,0 +1,3 @@
+export const isInteger = (value: any): boolean => {
+  return Number.isInteger(value);
+};