+++ packages/is.map/src/lib/isMap.ts
@@ -1,0 +1,3 @@
+export const isMap = (value: any): boolean => {
+  return value instanceof Map;
+};