+++ packages/is.object/src/lib/isObject.ts
@@ -1,0 +1,5 @@
+const objectConstructor = {}.constructor;
+
+export const isObject = (value: any): boolean => {
+  return typeof value === 'object' && value !== null && value.constructor === objectConstructor;
+};