+++ packages/is.regex/src/lib/isRegex.ts
@@ -1,0 +1,3 @@
+export const isRegex = (value: any): boolean => {
+  return value instanceof RegExp;
+};