+++ packages/is.set/src/lib/isSet.ts
@@ -1,0 +1,3 @@
+export const isSet = (value: any): boolean => {
+  return value instanceof Set;
+};