+++ packages/is.symbol/src/lib/isSymbol.ts
@@ -1,0 +1,3 @@
+export const isSymbol = (value: any): boolean => {
+  return typeof value === 'symbol';
+};