+++ packages/is.undefined/src/lib/isUndefined.ts
@@ -1,0 +1,3 @@
+export const isUndefined = (value: any): boolean => {
+  return typeof value === 'undefined';
+};