+++ packages/is.weakmap/src/lib/isWeakMap.ts
@@ -1,0 +1,3 @@
+export const isWeakMap = (value: any): boolean => {
+  return value instanceof WeakMap;
+};