+++ packages/is.weakset/src/lib/isWeakSet.ts
@@ -1,0 +1,3 @@
+export const isWeakSet = (value: any): boolean => {
+  return value instanceof WeakSet;
+};