+++ packages/is/src/index.spec.ts
@@ -1,0 +1,27 @@
+import * as guards from './index';
+
+describe('Guards barrel package', () => {
+  it('should export specific packages and interfaces', () => {
+    expect(Object.keys(guards)).toEqual([
+      'isArray',
+      'isBigint',
+      'isBoolean',
+      'isDate',
+      'isError',
+      'isFloat',
+      'isFunction',
+      'isInteger',
+      'isMap',
+      'isNull',
+      'isNumber',
+      'isObject',
+      'isRegex',
+      'isSet',
+      'isString',
+      'isSymbol',
+      'isUndefined',
+      'isWeakMap',
+      'isWeakSet',
+    ]);
+  });
+});