+++ packages/utils/src/index.ts
@@ -1,2 +1,3 @@
 export * from '@zerodep/errors';
 export * from '@zerodep/guards';
+export * from '@zerodep/is';