+++ packages/utils/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2022-04-25
+
+### Added
+
+- Boolean guard
+- Bigint guard
+- Date guard
+- Is utilities
+
 ## [0.1.0] - 2022-04-25
 
 ### Added